Fix syntax errors, warnings in string/stack

+++ MWSE/Stack.h
@@ -93,7 +93,7 @@ namespace mwse {
                 if (stack_top >= stack_size) {
                     stack_size += 32;
                     StackItem_t *new_stack = new StackItem_t[stack_size];
-                    for (int i = 0; i < stack_top; i++) {
+                    for (size_t i = 0; i < stack_top; i++) {
                         new_stack[i] = stack_storage[i];
                     }
                     delete [] stack_storage;
